customer-service/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
requests
PyMySQL
cryptography
pika
